fix(openapi): make API its own gradle subproject
If the openAPI will be out basis for provided endpoints, we will respectively need the YAMLs and generated artifacts in several places. Compiling everything in the model-server is quite a lot of code which is now isolated in its own subproject.
