+++ api/openapi/specifications/model-server-deprecated.yaml
