+++ api/openapi/specifications/model-server-light.yaml
