+++ api/openapi/specifications/model-server.yaml
@@ -553,12 +553,6 @@ paths:
           $ref: '#/components/responses/200'
         "404":
           $ref: '#/components/responses/404'
-  /metrics:
-    get:
-      operationId: getMetrics
-      responses:
-        "200":
-          $ref: '#/components/responses/200'
 components:
   responses:
     "200":
