+++ api/openapi/templates/data_class.mustache
@@ -1,0 +1,16 @@
+import kotlinx.serialization.Serializable
+
+/**
+ * {{{description}}}
+{{#vars}}
+ * @param {{{name}}} {{{description}}}
+{{/vars}}
+ */
+@Serializable
+data class {{classname}}(
+{{#requiredVars}}
+{{>data_class_req_var}}{{^-last}},
+{{/-last}}{{/requiredVars}}{{#hasRequired}}{{#hasOptional}},
+{{/hasOptional}}{{/hasRequired}}{{#optionalVars}}{{>data_class_opt_var}}{{^-last}},
+{{/-last}}{{/optionalVars}}
+)
