+++ model-client/build.gradle.kts
@@ -76,12 +76,12 @@ kotlin {
                 implementation(libs.apache.commons.lang)
                 implementation(libs.guava)
                 implementation(libs.apache.commons.io)
-                implementation("org.json:json:20231013")
+                implementation(libs.org.json)
                 implementation(libs.trove4j)
                 implementation(libs.apache.commons.collections)
 
-                implementation("com.google.oauth-client:google-oauth-client:1.34.1")
-                implementation("com.google.oauth-client:google-oauth-client-jetty:1.34.1")
+                implementation(libs.google.oauth.client)
+                implementation(libs.google.oauth.clientjetty)
 
                 implementation(libs.ktor.client.core)
                 implementation(libs.ktor.client.cio)
