+++ settings.gradle.kts
@@ -27,6 +27,7 @@ pluginManagement {
 
 rootProject.name = "modelix.core"
 
+include("api")
 include("authorization")
 include("bulk-model-sync-gradle")
 include("bulk-model-sync-lib")
